override `control` of standard gates where possible (#3631)
* move controlled gate class definitions into base gate module.

* fix cyclic imports

* update for new crx and cry

* linting

* linting

* linting

* minor

* wrong import of U3Gate

* module level pylint cyclic-import disable

* update mapper ground truth

* hack to get tox to work

* solve bug associated with ToffoliGate import.

* avoid multiple definitions

* update warning category.

* change stack level
