+++ qiskit/extensions/standard/ch.py
@@ -15,68 +15,9 @@
 """
 controlled-H gate.
 """
-import numpy as np
+import warnings
+# pylint: disable=unused-import
+from qiskit.extensions.standard.h import HGate, ch
 
-from qiskit.circuit import ControlledGate
-from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import QuantumRegister
-from qiskit.extensions.standard.h import HGate
-from qiskit.extensions.standard.cx import CnotGate
-from qiskit.extensions.standard.t import TGate, TdgGate
-from qiskit.extensions.standard.s import SGate, SdgGate
-
-
-class CHGate(ControlledGate):
-    """controlled-H gate."""
-
-    def __init__(self):
-        """Create new CH gate."""
-        super().__init__("ch", 2, [], num_ctrl_qubits=1)
-        self.base_gate = HGate
-        self.base_gate_name = "h"
-
-    def _define(self):
-        """
-        gate ch a,b {
-            s b;
-            h b;
-            t b;
-            cx a, b;
-            tdg b;
-            h b;
-            sdg b;
-        }
-        """
-        definition = []
-        q = QuantumRegister(2, "q")
-        rule = [
-            (SGate(), [q[1]], []),
-            (HGate(), [q[1]], []),
-            (TGate(), [q[1]], []),
-            (CnotGate(), [q[0], q[1]], []),
-            (TdgGate(), [q[1]], []),
-            (HGate(), [q[1]], []),
-            (SdgGate(), [q[1]], [])
-        ]
-        for inst in rule:
-            definition.append(inst)
-        self.definition = definition
-
-    def inverse(self):
-        """Invert this gate."""
-        return CHGate()  # self-inverse
-
-    def to_matrix(self):
-        """Return a Numpy.array for the Ch gate."""
-        return np.array([[1, 0, 0, 0],
-                         [0, 1/np.sqrt(2), 0, 1/np.sqrt(2)],
-                         [0, 0, 1, 0],
-                         [0, 1/np.sqrt(2), 0, -1/np.sqrt(2)]], dtype=complex)
-
-
-def ch(self, ctl, tgt):  # pylint: disable=invalid-name
-    """Apply CH from ctl to tgt."""
-    return self.append(CHGate(), [ctl, tgt], [])
-
-
-QuantumCircuit.ch = ch
+warnings.warn('This module is deprecated. The CHGate can now be found in h.py',
+              category=DeprecationWarning, stacklevel=2)
