+++ qiskit/extensions/standard/crz.py
@@ -15,50 +15,9 @@
 """
 controlled-rz gate.
 """
-from qiskit.circuit import ControlledGate
-from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import QuantumRegister
-from qiskit.extensions.standard.u1 import U1Gate
-from qiskit.extensions.standard.cx import CnotGate
-from qiskit.extensions.standard.rz import RZGate
+import warnings
+# pylint: disable=unused-import
+from qiskit.extensions.standard.rz import CrzGate, crz
 
-
-class CrzGate(ControlledGate):
-    """controlled-rz gate."""
-
-    def __init__(self, theta):
-        """Create new crz gate."""
-        super().__init__("crz", 2, [theta], num_ctrl_qubits=1)
-        self.base_gate = RZGate
-        self.base_gate_name = "rz"
-
-    def _define(self):
-        """
-        gate crz(lambda) a,b
-        { u1(lambda/2) b; cx a,b;
-          u1(-lambda/2) b; cx a,b;
-        }
-        """
-        definition = []
-        q = QuantumRegister(2, "q")
-        rule = [
-            (U1Gate(self.params[0] / 2), [q[1]], []),
-            (CnotGate(), [q[0], q[1]], []),
-            (U1Gate(-self.params[0] / 2), [q[1]], []),
-            (CnotGate(), [q[0], q[1]], [])
-        ]
-        for inst in rule:
-            definition.append(inst)
-        self.definition = definition
-
-    def inverse(self):
-        """Invert this gate."""
-        return CrzGate(-self.params[0])
-
-
-def crz(self, theta, ctl, tgt):
-    """Apply crz from ctl to tgt with angle theta."""
-    return self.append(CrzGate(theta), [ctl, tgt], [])
-
-
-QuantumCircuit.crz = crz
+warnings.warn('This module is deprecated. The CrzGate can now be found in rz.py',
+              category=DeprecationWarning, stacklevel=2)
