+++ qiskit/extensions/standard/cswap.py
@@ -15,51 +15,9 @@
 """
 Fredkin gate. Controlled-SWAP.
 """
-from qiskit.circuit import ControlledGate
-from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import QuantumRegister
-from qiskit.extensions.standard.cx import CnotGate
-from qiskit.extensions.standard.ccx import ToffoliGate
-from qiskit.extensions.standard.swap import SwapGate
+import warnings
+# pylint: disable=unused-import
+from qiskit.extensions.standard.swap import FredkinGate, cswap
 
-
-class FredkinGate(ControlledGate):
-    """Fredkin gate."""
-
-    def __init__(self):
-        """Create new Fredkin gate."""
-        super().__init__("cswap", 3, [], num_ctrl_qubits=1)
-        self.base_gate = SwapGate
-        self.base_gate_name = "swap"
-
-    def _define(self):
-        """
-        gate cswap a,b,c
-        { cx c,b;
-          ccx a,b,c;
-          cx c,b;
-        }
-        """
-        definition = []
-        q = QuantumRegister(3, "q")
-        rule = [
-            (CnotGate(), [q[2], q[1]], []),
-            (ToffoliGate(), [q[0], q[1], q[2]], []),
-            (CnotGate(), [q[2], q[1]], [])
-        ]
-        for inst in rule:
-            definition.append(inst)
-        self.definition = definition
-
-    def inverse(self):
-        """Invert this gate."""
-        return FredkinGate()  # self-inverse
-
-
-def cswap(self, ctl, tgt1, tgt2):
-    """Apply Fredkin to circuit."""
-    return self.append(FredkinGate(), [ctl, tgt1, tgt2], [])
-
-
-QuantumCircuit.cswap = cswap
-QuantumCircuit.fredkin = cswap
+warnings.warn('This module is deprecated. The FredkinGate can now be found in swap.py',
+              category=DeprecationWarning, stacklevel=2)
