+++ qiskit/extensions/standard/cu1.py
@@ -15,51 +15,9 @@
 """
 controlled-u1 gate.
 """
-from qiskit.circuit import ControlledGate
-from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import QuantumRegister
-from qiskit.extensions.standard.u1 import U1Gate
-from qiskit.extensions.standard.cx import CnotGate
+import warnings
+# pylint: disable=unused-import
+from qiskit.extensions.standard.u1 import Cu1Gate, cu1
 
-
-class Cu1Gate(ControlledGate):
-    """controlled-u1 gate."""
-
-    def __init__(self, theta):
-        """Create new cu1 gate."""
-        super().__init__("cu1", 2, [theta], num_ctrl_qubits=1)
-        self.base_gate = U1Gate
-        self.base_gate_name = "u1"
-
-    def _define(self):
-        """
-        gate cu1(lambda) a,b
-        { u1(lambda/2) a; cx a,b;
-          u1(-lambda/2) b; cx a,b;
-          u1(lambda/2) b;
-        }
-        """
-        definition = []
-        q = QuantumRegister(2, "q")
-        rule = [
-            (U1Gate(self.params[0] / 2), [q[0]], []),
-            (CnotGate(), [q[0], q[1]], []),
-            (U1Gate(-self.params[0] / 2), [q[1]], []),
-            (CnotGate(), [q[0], q[1]], []),
-            (U1Gate(self.params[0] / 2), [q[1]], [])
-        ]
-        for inst in rule:
-            definition.append(inst)
-        self.definition = definition
-
-    def inverse(self):
-        """Invert this gate."""
-        return Cu1Gate(-self.params[0])
-
-
-def cu1(self, theta, ctl, tgt):
-    """Apply cu1 from ctl to tgt with angle theta."""
-    return self.append(Cu1Gate(theta), [ctl, tgt], [])
-
-
-QuantumCircuit.cu1 = cu1
+warnings.warn('This module is deprecated. The Cu1Gate can now be found in u1.py',
+              category=DeprecationWarning, stacklevel=2)
