+++ qiskit/extensions/standard/cz.py
@@ -15,55 +15,9 @@
 """
 controlled-Phase gate.
 """
-import numpy
+import warnings
+# pylint: disable=unused-import
+from qiskit.extensions.standard.z import CzGate, cz
 
-from qiskit.circuit import ControlledGate
-from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import QuantumRegister
-from qiskit.extensions.standard.h import HGate
-from qiskit.extensions.standard.cx import CnotGate
-from qiskit.extensions.standard.z import ZGate
-
-
-class CzGate(ControlledGate):
-    """controlled-Z gate."""
-
-    def __init__(self, label=None):
-        """Create new CZ gate."""
-        super().__init__("cz", 2, [], label=label, num_ctrl_qubits=1)
-        self.base_gate = ZGate
-        self.base_gate_name = "z"
-
-    def _define(self):
-        """
-        gate cz a,b { h b; cx a,b; h b; }
-        """
-        definition = []
-        q = QuantumRegister(2, "q")
-        rule = [
-            (HGate(), [q[1]], []),
-            (CnotGate(), [q[0], q[1]], []),
-            (HGate(), [q[1]], [])
-        ]
-        for inst in rule:
-            definition.append(inst)
-        self.definition = definition
-
-    def inverse(self):
-        """Invert this gate."""
-        return CzGate()  # self-inverse
-
-    def to_matrix(self):
-        """Return a Numpy.array for the Cz gate."""
-        return numpy.array([[1, 0, 0, 0],
-                            [0, 1, 0, 0],
-                            [0, 0, 1, 0],
-                            [0, 0, 0, -1]], dtype=complex)
-
-
-def cz(self, ctl, tgt):  # pylint: disable=invalid-name
-    """Apply CZ to circuit."""
-    return self.append(CzGate(), [ctl, tgt], [])
-
-
-QuantumCircuit.cz = cz
+warnings.warn('This module is deprecated. The CzGate can now be found in z.py',
+              category=DeprecationWarning, stacklevel=2)
