+++ qiskit/extensions/standard/rx.py
@@ -17,10 +17,11 @@ Rotation around the x-axis.
 """
 import math
 import numpy
+from qiskit.circuit import ControlledGate
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
 from qiskit.circuit import QuantumRegister
-from qiskit.extensions.standard.r import RGate
+from qiskit.qasm import pi
 
 
 class RXGate(Gate):
@@ -34,6 +35,7 @@ class RXGate(Gate):
         """
         gate rx(theta) a {r(theta, 0) a;}
         """
+        from qiskit.extensions.standard.r import RGate
         definition = []
         q = QuantumRegister(1, "q")
         rule = [
@@ -42,6 +44,20 @@ class RXGate(Gate):
         for inst in rule:
             definition.append(inst)
         self.definition = definition
+
+    def control(self, num_ctrl_qubits=1, label=None):
+        """Controlled version of this gate.
+
+        Args:
+            num_ctrl_qubits (int): number of control qubits.
+            label (str or None): An optional label for the gate [Default: None]
+
+        Returns:
+            ControlledGate: controlled version of this gate.
+        """
+        if num_ctrl_qubits == 1:
+            return CrxGate(self.params[0])
+        return super().control(num_ctrl_qubits=num_ctrl_qubits, label=label)
 
     def inverse(self):
         """Invert this gate.
@@ -64,3 +80,51 @@ def rx(self, theta, q):  # pylint: disable=invalid-name
 
 
 QuantumCircuit.rx = rx
+
+
+class CrxGate(ControlledGate):
+    """controlled-rx gate."""
+
+    def __init__(self, theta):
+        """Create new crx gate."""
+        super().__init__('crx', 2, [theta], num_ctrl_qubits=1)
+        self.base_gate = RXGate
+        self.base_gate_name = 'rx'
+
+    def _define(self):
+        """
+        gate cu3(theta,phi,lambda) c, t
+        { u1(pi/2) t;
+          cx c,t;
+          u3(-theta/2,0,0) t;
+          cx c,t;
+          u3(theta/2,-pi/2,0) t;
+        }
+        """
+        from qiskit.extensions.standard.u1 import U1Gate
+        from qiskit.extensions.standard.u3 import U3Gate
+        from qiskit.extensions.standard.x import CnotGate
+        definition = []
+        q = QuantumRegister(2, 'q')
+        rule = [
+            (U1Gate(pi / 2), [q[1]], []),
+            (CnotGate(), [q[0], q[1]], []),
+            (U3Gate(-self.params[0] / 2, 0, 0), [q[1]], []),
+            (CnotGate(), [q[0], q[1]], []),
+            (U3Gate(self.params[0] / 2, -pi / 2, 0), [q[1]], [])
+        ]
+        for inst in rule:
+            definition.append(inst)
+        self.definition = definition
+
+    def inverse(self):
+        """Invert this gate."""
+        return CrxGate(-self.params[0])
+
+
+def crx(self, theta, ctl, tgt):
+    """Apply crx from ctl to tgt with angle theta."""
+    return self.append(CrxGate(theta), [ctl, tgt], [])
+
+
+QuantumCircuit.crx = crx
