+++ qiskit/extensions/standard/ry.py
@@ -17,11 +17,11 @@ Rotation around the y-axis.
 """
 import math
 import numpy
+from qiskit.circuit import ControlledGate
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
 from qiskit.circuit import QuantumRegister
 from qiskit.qasm import pi
-from qiskit.extensions.standard.r import RGate
 
 
 class RYGate(Gate):
@@ -35,6 +35,7 @@ class RYGate(Gate):
         """
         gate ry(theta) a { r(theta, pi/2) a; }
         """
+        from qiskit.extensions.standard.r import RGate
         definition = []
         q = QuantumRegister(1, "q")
         rule = [
@@ -43,6 +44,20 @@ class RYGate(Gate):
         for inst in rule:
             definition.append(inst)
         self.definition = definition
+
+    def control(self, num_ctrl_qubits=1, label=None):
+        """Controlled version of this gate.
+
+        Args:
+            num_ctrl_qubits (int): number of control qubits.
+            label (str or None): An optional label for the gate [Default: None]
+
+        Returns:
+            ControlledGate: controlled version of this gate.
+        """
+        if num_ctrl_qubits == 1:
+            return CryGate(self.params[0])
+        return super().control(num_ctrl_qubits=num_ctrl_qubits, label=label)
 
     def inverse(self):
         """Invert this gate.
@@ -65,3 +80,47 @@ def ry(self, theta, q):  # pylint: disable=invalid-name
 
 
 QuantumCircuit.ry = ry
+
+
+class CryGate(ControlledGate):
+    """controlled-ry gate."""
+
+    def __init__(self, theta):
+        """Create new cry gate."""
+        super().__init__("cry", 2, [theta], num_ctrl_qubits=1)
+        self.base_gate = RYGate
+        self.base_gate_name = "ry"
+
+    def _define(self):
+        """
+        gate cry(lambda) a,b
+        { u3(lambda/2,0,0) b; cx a,b;
+          u3(-lambda/2,0,0) b; cx a,b;
+        }
+
+        """
+        from qiskit.extensions.standard.x import CnotGate
+        from qiskit.extensions.standard.u3 import U3Gate
+        definition = []
+        q = QuantumRegister(2, "q")
+        rule = [
+            (U3Gate(self.params[0] / 2, 0, 0), [q[1]], []),
+            (CnotGate(), [q[0], q[1]], []),
+            (U3Gate(-self.params[0] / 2, 0, 0), [q[1]], []),
+            (CnotGate(), [q[0], q[1]], [])
+        ]
+        for inst in rule:
+            definition.append(inst)
+        self.definition = definition
+
+    def inverse(self):
+        """Invert this gate."""
+        return CryGate(-self.params[0])
+
+
+def cry(self, theta, ctl, tgt):
+    """Apply cry from ctl to tgt with angle theta."""
+    return self.append(CryGate(theta), [ctl, tgt], [])
+
+
+QuantumCircuit.cry = cry
