+++ qiskit/extensions/standard/x.py
@@ -16,12 +16,11 @@
 Pauli X (bit-flip) gate.
 """
 import numpy
-
+from qiskit.circuit import ControlledGate
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
 from qiskit.circuit import QuantumRegister
 from qiskit.qasm import pi
-from qiskit.extensions.standard.u3 import U3Gate
 
 
 class XGate(Gate):
@@ -37,6 +36,7 @@ class XGate(Gate):
         u3(pi,0,pi) a;
         }
         """
+        from qiskit.extensions.standard.u3 import U3Gate
         definition = []
         q = QuantumRegister(1, "q")
         rule = [
@@ -45,6 +45,22 @@ class XGate(Gate):
         for inst in rule:
             definition.append(inst)
         self.definition = definition
+
+    def control(self, num_ctrl_qubits=1, label=None):
+        """Controlled version of this gate.
+
+        Args:
+            num_ctrl_qubits (int): number of control qubits.
+            label (str or None): An optional label for the gate [Default: None]
+
+        Returns:
+            ControlledGate: controlled version of this gate.
+        """
+        if num_ctrl_qubits == 1:
+            return CnotGate()
+        elif num_ctrl_qubits == 2:
+            return ToffoliGate()
+        return super().control(num_ctrl_qubits=num_ctrl_qubits, label=label)
 
     def inverse(self):
         """Invert this gate."""
@@ -62,3 +78,116 @@ def x(self, q):
 
 
 QuantumCircuit.x = x
+
+
+class CnotGate(ControlledGate):
+    """controlled-NOT gate."""
+
+    def __init__(self):
+        """Create new CNOT gate."""
+        super().__init__("cx", 2, [], num_ctrl_qubits=1)
+        self.base_gate = XGate
+        self.base_gate_name = "x"
+
+    def control(self, num_ctrl_qubits=1, label=None):
+        """Controlled version of this gate.
+
+        Args:
+            num_ctrl_qubits (int): number of control qubits.
+            label (str or None): An optional label for the gate [Default: None]
+
+        Returns:
+            ControlledGate: controlled version of this gate.
+        """
+        if num_ctrl_qubits == 1:
+            return ToffoliGate()
+        return super().control(num_ctrl_qubits=num_ctrl_qubits, label=label)
+
+    def inverse(self):
+        """Invert this gate."""
+        return CnotGate()  # self-inverse
+
+    def to_matrix(self):
+        """Return a Numpy.array for the Cx gate."""
+        return numpy.array([[1, 0, 0, 0],
+                            [0, 0, 0, 1],
+                            [0, 0, 1, 0],
+                            [0, 1, 0, 0]], dtype=complex)
+
+
+def cx(self, ctl, tgt):  # pylint: disable=invalid-name
+    """Apply CX from ctl to tgt."""
+    return self.append(CnotGate(), [ctl, tgt], [])
+
+
+QuantumCircuit.cx = cx
+QuantumCircuit.cnot = cx
+
+
+class ToffoliGate(ControlledGate):
+    """Toffoli gate."""
+
+    def __init__(self):
+        """Create new Toffoli gate."""
+        super().__init__("ccx", 3, [], num_ctrl_qubits=2)
+        self.base_gate = XGate
+        self.base_gate_name = "x"
+
+    def _define(self):
+        """
+        gate ccx a,b,c
+        {
+        h c; cx b,c; tdg c; cx a,c;
+        t c; cx b,c; tdg c; cx a,c;
+        t b; t c; h c; cx a,b;
+        t a; tdg b; cx a,b;}
+        """
+        from qiskit.extensions.standard.h import HGate
+        from qiskit.extensions.standard.t import TGate
+        from qiskit.extensions.standard.t import TdgGate
+        definition = []
+        q = QuantumRegister(3, "q")
+        rule = [
+            (HGate(), [q[2]], []),
+            (CnotGate(), [q[1], q[2]], []),
+            (TdgGate(), [q[2]], []),
+            (CnotGate(), [q[0], q[2]], []),
+            (TGate(), [q[2]], []),
+            (CnotGate(), [q[1], q[2]], []),
+            (TdgGate(), [q[2]], []),
+            (CnotGate(), [q[0], q[2]], []),
+            (TGate(), [q[1]], []),
+            (TGate(), [q[2]], []),
+            (HGate(), [q[2]], []),
+            (CnotGate(), [q[0], q[1]], []),
+            (TGate(), [q[0]], []),
+            (TdgGate(), [q[1]], []),
+            (CnotGate(), [q[0], q[1]], [])
+        ]
+        for inst in rule:
+            definition.append(inst)
+        self.definition = definition
+
+    def inverse(self):
+        """Invert this gate."""
+        return ToffoliGate()  # self-inverse
+
+    def to_matrix(self):
+        """Return a Numpy.array for the Toffoli gate."""
+        return numpy.array([[1, 0, 0, 0, 0, 0, 0, 0],
+                            [0, 1, 0, 0, 0, 0, 0, 0],
+                            [0, 0, 1, 0, 0, 0, 0, 0],
+                            [0, 0, 0, 0, 0, 0, 0, 1],
+                            [0, 0, 0, 0, 1, 0, 0, 0],
+                            [0, 0, 0, 0, 0, 1, 0, 0],
+                            [0, 0, 0, 0, 0, 0, 1, 0],
+                            [0, 0, 0, 1, 0, 0, 0, 0]], dtype=complex)
+
+
+def ccx(self, ctl1, ctl2, tgt):
+    """Apply Toffoli to ctl1 and ctl2 to tgt."""
+    return self.append(ToffoliGate(), [ctl1, ctl2, tgt], [])
+
+
+QuantumCircuit.ccx = ccx
+QuantumCircuit.toffoli = ccx
