+++ qiskit/extensions/standard/y.py
@@ -17,10 +17,10 @@ Pauli Y (bit-phase-flip) gate.
 """
 import numpy
 from qiskit.circuit import Gate
+from qiskit.circuit import ControlledGate
 from qiskit.circuit import QuantumRegister
 from qiskit.circuit import QuantumCircuit
 from qiskit.qasm import pi
-from qiskit.extensions.standard.u3 import U3Gate
 
 
 class YGate(Gate):
@@ -31,6 +31,7 @@ class YGate(Gate):
         super().__init__("y", 1, [], label=label)
 
     def _define(self):
+        from qiskit.extensions.standard.u3 import U3Gate
         definition = []
         q = QuantumRegister(1, "q")
         rule = [
@@ -39,6 +40,20 @@ class YGate(Gate):
         for inst in rule:
             definition.append(inst)
         self.definition = definition
+
+    def control(self, num_ctrl_qubits=1, label=None):
+        """Controlled version of this gate.
+
+        Args:
+            num_ctrl_qubits (int): number of control qubits.
+            label (str or None): An optional label for the gate [Default: None]
+
+        Returns:
+            ControlledGate: controlled version of this gate.
+        """
+        if num_ctrl_qubits == 1:
+            return CyGate()
+        return super().control(num_ctrl_qubits=num_ctrl_qubits, label=label)
 
     def inverse(self):
         """Invert this gate."""
@@ -56,3 +71,43 @@ def y(self, q):
 
 
 QuantumCircuit.y = y
+
+
+class CyGate(ControlledGate):
+    """controlled-Y gate."""
+
+    def __init__(self):
+        """Create new CY gate."""
+        super().__init__("cy", 2, [], num_ctrl_qubits=1)
+        self.base_gate = YGate
+        self.base_gate_name = "y"
+
+    def _define(self):
+        """
+        gate cy a,b { sdg b; cx a,b; s b; }
+        """
+        from qiskit.extensions.standard.s import SGate
+        from qiskit.extensions.standard.s import SdgGate
+        from qiskit.extensions.standard.x import CnotGate
+        definition = []
+        q = QuantumRegister(2, "q")
+        rule = [
+            (SdgGate(), [q[1]], []),
+            (CnotGate(), [q[0], q[1]], []),
+            (SGate(), [q[1]], [])
+        ]
+        for inst in rule:
+            definition.append(inst)
+        self.definition = definition
+
+    def inverse(self):
+        """Invert this gate."""
+        return CyGate()  # self-inverse
+
+
+def cy(self, ctl, tgt):  # pylint: disable=invalid-name
+    """Apply CY to circuit."""
+    return self.append(CyGate(), [ctl, tgt], [])
+
+
+QuantumCircuit.cy = cy
