+++ qiskit/extensions/standard/z.py
@@ -17,10 +17,10 @@ Pauli Z (phase-flip) gate.
 """
 import numpy
 from qiskit.circuit import Gate
+from qiskit.circuit import ControlledGate
 from qiskit.circuit import QuantumCircuit
 from qiskit.circuit import QuantumRegister
 from qiskit.qasm import pi
-from qiskit.extensions.standard.u1 import U1Gate
 
 
 class ZGate(Gate):
@@ -31,6 +31,7 @@ class ZGate(Gate):
         super().__init__("z", 1, [], label=label)
 
     def _define(self):
+        from qiskit.extensions.standard.u1 import U1Gate
         definition = []
         q = QuantumRegister(1, "q")
         rule = [
@@ -39,6 +40,20 @@ class ZGate(Gate):
         for inst in rule:
             definition.append(inst)
         self.definition = definition
+
+    def control(self, num_ctrl_qubits=1, label=None):
+        """Controlled version of this gate.
+
+        Args:
+            num_ctrl_qubits (int): number of control qubits.
+            label (str or None): An optional label for the gate [Default: None]
+
+        Returns:
+            ControlledGate: controlled version of this gate.
+        """
+        if num_ctrl_qubits == 1:
+            return CzGate()
+        return super().control(num_ctrl_qubits=num_ctrl_qubits, label=label)
 
     def inverse(self):
         """Invert this gate."""
@@ -56,3 +71,49 @@ def z(self, q):
 
 
 QuantumCircuit.z = z
+
+
+class CzGate(ControlledGate):
+    """controlled-Z gate."""
+
+    def __init__(self, label=None):
+        """Create new CZ gate."""
+        super().__init__("cz", 2, [], label=label, num_ctrl_qubits=1)
+        self.base_gate = ZGate
+        self.base_gate_name = "z"
+
+    def _define(self):
+        """
+        gate cz a,b { h b; cx a,b; h b; }
+        """
+        from qiskit.extensions.standard.h import HGate
+        from qiskit.extensions.standard.x import CnotGate
+        definition = []
+        q = QuantumRegister(2, "q")
+        rule = [
+            (HGate(), [q[1]], []),
+            (CnotGate(), [q[0], q[1]], []),
+            (HGate(), [q[1]], [])
+        ]
+        for inst in rule:
+            definition.append(inst)
+        self.definition = definition
+
+    def inverse(self):
+        """Invert this gate."""
+        return CzGate()  # self-inverse
+
+    def to_matrix(self):
+        """Return a Numpy.array for the Cz gate."""
+        return numpy.array([[1, 0, 0, 0],
+                            [0, 1, 0, 0],
+                            [0, 0, 1, 0],
+                            [0, 0, 0, -1]], dtype=complex)
+
+
+def cz(self, ctl, tgt):  # pylint: disable=invalid-name
+    """Apply CZ to circuit."""
+    return self.append(CzGate(), [ctl, tgt], [])
+
+
+QuantumCircuit.cz = cz
